requirements.txt
flask==3.0.2
requests==2.32.3
beautifulsoup4==4.12.3
pycryptodome==3.20.0
gunicorn==22.0.0